fixed create user function to ignore duplicates, added delete user function, fixed store todo to store into user todo array and edited all other todo functions

+++ backend/src/todos/removeTodo.js
@@ -14,9 +14,8 @@ const dbName = "mydb";
 export async function removeTodo(userId, todo) {
   await client.connect();
   const db = client.db(dbName);
-  const userID = userId.toString();
-  const collection = db.collection(userID);
-  const insertResult = await collection.deleteOne({ todo: todo });
+  const collection = db.collection("users");
+  const insertResult = await collection.deleteOne({ _id: userId, todo: todo });
   console.log("Removed documents =>", insertResult);
   return "done.";
 }
